requirements.txt
python_bcrypt==0.3.2
bcrypt==3.1.7
Django==3.1
django-allauth==0.42.0
django-bootstrap-form==3.4
django-cors-headers==3.4.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-email-verification==0.0.5.post1
django-filter==2.3.0
django-multiupload==0.5.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
future==0.18.2
Jinja2==2.11.2
psycopg2==2.8.5
pycodestyle==2.6.0
requests==2.21.0
requests-oauthlib==1.2.0
six==1.15.0
twilio==6.44.2
oauthlib==3.0.1
